+++ source/dotnet/Services/WebApi.IntegrationTests/WebApi/V3/ProcessStepBalanceResponsiblePartyTests.cs
@@ -19,7 +19,7 @@ using Energinet.DataHub.Wholesale.Application.ProcessStep;
 using Energinet.DataHub.Wholesale.Contracts;
 using Energinet.DataHub.Wholesale.WebApi.IntegrationTests.Fixtures.TestCommon.Fixture.WebApi;
 using Energinet.DataHub.Wholesale.WebApi.IntegrationTests.Fixtures.WebApi;
-using Energinet.DataHub.Wholesale.WebApi.V3.ProcessStepActor;
+using Energinet.DataHub.Wholesale.WebApi.V3;
 using FluentAssertions;
 using Moq;
 using Xunit;
