+++ source/dotnet/Services/WebApi/V2/ProcessStepControllerV23.cs
@@ -17,7 +17,7 @@ using Energinet.DataHub.Wholesale.Contracts;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Energinet.DataHub.Wholesale.WebApi.Controllers.V2;
+namespace Energinet.DataHub.Wholesale.WebApi.V2;
 
 [ApiController]
 [Route("v2.3/ProcessStepResult")]
