+++ source/dotnet/Services/WebApi/V3/ProcessStepBalanceResponsibleParty/ProcessStepBalanceResponsiblePartyController.cs
@@ -14,7 +14,6 @@
 
 using Energinet.DataHub.Wholesale.Application.ProcessStep;
 using Energinet.DataHub.Wholesale.Contracts;
-using Energinet.DataHub.Wholesale.WebApi.V3.ProcessStepActor;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
